requirements.txt
discord.py==2.3.2
pyppeteer
firebase-admin
pytz
pytesseract
webdriver-manager
Pillow
